rm StorageTransaction trait
