+++ minigu/storage/src/lib.rs
@@ -6,11 +6,8 @@ pub mod error;
 pub mod iterators;
 pub mod memory;
 pub mod model;
-pub mod storage;
-pub mod transaction;
+
 pub mod wal;
 pub use error::StorageResult;
 pub use memory::memory_graph::MemoryGraph;
-pub use memory::transaction::{MemTransaction, TransactionHandle};
-pub use storage::StorageTransaction;
-pub use transaction::IsolationLevel;
+pub use memory::transaction::{IsolationLevel, MemTransaction, TransactionHandle};
